pyproject.toml
[tool.poetry]
name = "skypro-libraries"
version = "0.1.0"
description = ""
authors = ["Mikhail Belin <belinma@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.lint.dependencies]
black = "^24.8.0"
mypy = "^1.11.2"
flake8 = "^7.1.1"
isort = "^5.13.2"


[tool.poetry.group.pandas.dependencies]


[tool.poetry.group.dev.dependencies]
pandas = "^2.2.2"
openpyxl = "^3.1.5"
requests = "^2.32.3"
pytest = "^8.3.2"
logging = "^0.4.9.6"
python-dotenv = "^1.0.1"
pandas-stubs = "^2.2.2.240807"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = '.venv'

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.git
  )/
)
'''

[tool.isort]
line_length = 119